Reorganised code.
